renamed addMex to invertMex, implemented a driver for the mex function, removed the input validation on the C++ side
