--- cpp/addMex/invertMex.cpp
@@ -1,51 +1,0 @@
-#include "mex.hpp"
-#include "mexAdapter.hpp"
-
-using namespace matlab::data;
-using matlab::mex::ArgumentList;
-
-class MexFunction : public matlab::mex::Function {
-public:
-    void operator()(ArgumentList outputs, ArgumentList inputs) {
-        checkArguments(outputs, inputs);
-        const double offSet = inputs[0][0];
-        TypedArray<double> doubleArray = std::move(inputs[1]);
-        for (auto& elem : doubleArray) {
-            elem += offSet;
-        }
-        outputs[0] = doubleArray;
-    }
-
-    void checkArguments(ArgumentList outputs, ArgumentList inputs) {
-        // Get pointer to engine
-        std::shared_ptr<matlab::engine::MATLABEngine> matlabPtr = getEngine();
-
-        // Get array factory
-        ArrayFactory factory;
-
-        // Check offset argument: First input must be scalar double
-        if (inputs[0].getType() != ArrayType::DOUBLE ||
-            inputs[0].getType() == ArrayType::COMPLEX_DOUBLE ||
-            inputs[0].getNumberOfElements() != 1)
-        {
-            matlabPtr->feval(u"error",
-                0,
-                std::vector<Array>({ factory.createScalar("First input must be scalar double") }));
-        }
-
-        // Check array argument: Second input must be double array
-        if (inputs[1].getType() != ArrayType::DOUBLE ||
-            inputs[1].getType() == ArrayType::COMPLEX_DOUBLE)
-        {
-            matlabPtr->feval(u"error",
-                0,
-                std::vector<Array>({ factory.createScalar("Input must be double array") }));
-        }
-        // Check number of outputs
-        if (outputs.size() > 1) {
-            matlabPtr->feval(u"error",
-                0,
-                std::vector<Array>({ factory.createScalar("Only one output is returned") }));
-        }
-    }
-};
